Move tag lifecycle out of reading into the tagging module
Tag and tag-group use cases, the repository protocol, the ORM models,
mappers, schemas and the router all move under application/tagging and
infrastructure/tagging, wired through a new TaggingContainer. Routes,
paths, schemas and generated OpenAPI are byte-for-byte unchanged, as are
the tables.

Associations stay with the module that owns them: add_tag_to_highlight
and remove_tag_from_highlight move off TagRepositoryProtocol onto
HighlightRepositoryProtocol and into HighlightRepository, and the
highlight_tags table stays in reading (note_tags stays in notes). The
tagging repository still reads both link tables to tell which tags are in
active use, and Tag keeps its highlights relationship so deleting a tag
clears its association rows exactly as before.
