+++ backend/src/application/notes/use_cases/get_note_use_case.py
@@ -6,7 +6,7 @@ from src.application.notes.protocols.note_repository import NoteRepositoryProtoc
 from src.application.notes.use_cases.dtos import NoteWithLinkedEntities
 from src.application.notes.use_cases.helpers import hydrate_note_links
 from src.application.reading.protocols.highlight_repository import HighlightRepositoryProtocol
-from src.application.reading.protocols.tag_repository import (
+from src.application.tagging.protocols.tag_repository import (
     TagRepositoryProtocol,
 )
 from src.domain.common.value_objects import NoteId, UserId
