+++ backend/src/application/tagging/__init__.py
@@ -1,0 +1,1 @@
+"""Tagging application context."""
