+++ backend/src/application/tagging/protocols/__init__.py
@@ -1,0 +1,5 @@
+"""Tagging application protocols."""
+
+from src.application.tagging.protocols.tag_repository import TagRepositoryProtocol
+
+__all__ = ["TagRepositoryProtocol"]
