+++ backend/src/application/tagging/protocols/tag_repository.py
@@ -4,7 +4,6 @@ from typing import Protocol
 
 from src.domain.common.value_objects.ids import (
     BookId,
-    HighlightId,
     TagGroupId,
     TagId,
     UserId,
@@ -45,13 +44,3 @@ class TagRepositoryProtocol(Protocol):
     async def delete_group(self, group_id: TagGroupId) -> bool: ...
 
     async def check_group_exists(self, group_id: TagGroupId) -> bool: ...
-
-    # Tag-Highlight association methods
-
-    async def add_tag_to_highlight(
-        self, highlight_id: HighlightId, tag_id: TagId, user_id: UserId
-    ) -> bool: ...
-
-    async def remove_tag_from_highlight(
-        self, highlight_id: HighlightId, tag_id: TagId, user_id: UserId
-    ) -> bool: ...
