+++ backend/src/application/tagging/use_cases/__init__.py
@@ -1,0 +1,1 @@
+"""Tagging use cases."""
