+++ backend/src/application/tagging/use_cases/tag_groups/delete_tag_group_use_case.py
@@ -4,7 +4,7 @@ Use case for deleting a tag group.
 
 import structlog
 
-from src.application.reading.protocols.tag_repository import (
+from src.application.tagging.protocols.tag_repository import (
     TagRepositoryProtocol,
 )
 from src.domain.common.value_objects.ids import TagGroupId
