+++ backend/src/application/tagging/use_cases/tag_groups/update_tag_group_association_use_case.py
@@ -5,7 +5,7 @@ Use case for updating a tag's group association.
 import structlog
 
 from src.application.common.ownership import require_belongs_to_book
-from src.application.reading.protocols.tag_repository import (
+from src.application.tagging.protocols.tag_repository import (
     TagRepositoryProtocol,
 )
 from src.domain.common.value_objects.ids import (
