+++ backend/src/application/tagging/use_cases/tags/get_tags_for_book_use_case.py
@@ -2,7 +2,7 @@
 
 from src.application.common.ownership import require_book
 from src.application.library.protocols.book_repository import BookRepositoryProtocol
-from src.application.reading.protocols.tag_repository import (
+from src.application.tagging.protocols.tag_repository import (
     TagRepositoryProtocol,
 )
 from src.domain.common.value_objects.ids import BookId, UserId
