+++ backend/src/infrastructure/reading/mappers/__init__.py
@@ -2,14 +2,10 @@ from .bookmark_mapper import BookmarkMapper
 from .highlight_mapper import HighlightMapper
 from .highlight_style_mapper import HighlightStyleMapper
 from .reading_session_mapper import ReadingSessionMapper
-from .tag_group_mapper import TagGroupMapper
-from .tag_mapper import TagMapper
 
 __all__ = [
     "BookmarkMapper",
     "HighlightMapper",
     "HighlightStyleMapper",
     "ReadingSessionMapper",
-    "TagGroupMapper",
-    "TagMapper",
 ]
