+++ backend/src/infrastructure/reading/orm/highlight_model.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     from src.infrastructure.reading.orm.bookmark_model import Bookmark
     from src.infrastructure.reading.orm.highlight_style_model import HighlightStyle
     from src.infrastructure.reading.orm.reading_session_model import ReadingSession
-    from src.infrastructure.reading.orm.tag_model import Tag
+    from src.infrastructure.tagging.orm.tag_model import Tag
 
 
 class Highlight(Base):
