+++ backend/src/infrastructure/reading/repositories/__init__.py
@@ -2,12 +2,10 @@ from .bookmark_repository import BookmarkRepository
 from .highlight_repository import HighlightRepository
 from .highlight_style_repository import HighlightStyleRepository
 from .reading_session_repository import ReadingSessionRepository
-from .tag_repository import TagRepository
 
 __all__ = [
     "BookmarkRepository",
     "HighlightRepository",
     "HighlightStyleRepository",
     "ReadingSessionRepository",
-    "TagRepository",
 ]
