+++ backend/src/infrastructure/tagging/__init__.py
@@ -1,0 +1,1 @@
+"""Tagging infrastructure context."""
