+++ backend/src/infrastructure/tagging/mappers/__init__.py
@@ -1,0 +1,6 @@
+"""Tagging mappers."""
+
+from src.infrastructure.tagging.mappers.tag_group_mapper import TagGroupMapper
+from src.infrastructure.tagging.mappers.tag_mapper import TagMapper
+
+__all__ = ["TagGroupMapper", "TagMapper"]
