+++ backend/src/infrastructure/tagging/orm/tag_group_model.py
@@ -10,7 +10,7 @@ from src.database import Base
 
 if TYPE_CHECKING:
     from src.infrastructure.library.orm.book_model import Book
-    from src.infrastructure.reading.orm.tag_model import Tag
+    from src.infrastructure.tagging.orm.tag_model import Tag
 
 
 class TagGroup(Base):
