+++ backend/src/infrastructure/tagging/repositories/__init__.py
@@ -1,0 +1,5 @@
+"""Tagging repositories."""
+
+from src.infrastructure.tagging.repositories.tag_repository import TagRepository
+
+__all__ = ["TagRepository"]
