+++ backend/src/infrastructure/tagging/routers/__init__.py
@@ -1,0 +1,1 @@
+"""Tagging routers."""
