+++ backend/src/infrastructure/tagging/schemas/__init__.py
@@ -1,0 +1,23 @@
+"""Tagging API schemas."""
+
+from src.infrastructure.tagging.schemas.tag_schemas import (
+    Tag,
+    TagAssociationRequest,
+    TagCreateRequest,
+    TagGroup,
+    TagGroupCreateRequest,
+    TagGroupInBook,
+    TagInBook,
+    TagUpdateRequest,
+)
+
+__all__ = [
+    "Tag",
+    "TagAssociationRequest",
+    "TagCreateRequest",
+    "TagGroup",
+    "TagGroupCreateRequest",
+    "TagGroupInBook",
+    "TagInBook",
+    "TagUpdateRequest",
+]
